renamed transaction.rb
