+++ lib/bitstampede/Mapper.rb
@@ -1,6 +1,6 @@
 require_relative './entities/UserBalance'
 require_relative './entities/Order'
-require_relative './entities/transaction'
+require_relative './entities/Transaction'
 require_relative './entities/Ticker'
 
 module Bitstampede
