+++ lib/bitstampede/entities/Transaction.rb
